pyproject.toml
[tool.poetry]
name = "noisemon"
version = "1.0.0"
description = "Entity Linking Service"
authors = ["Max Ermakov <maksim_ermakov@protonmail.com>"]
packages = [{include = "noisemon"}]

[tool.poetry.dependencies]
python = ">=3.12"
regex = "^2021.7.6"
datasets = "^2.13.1"
einops = "^0.6.1"
pgvector = "^0.2.3"
pydantic-settings = {extras = ["dotenv"], version = "^2.5.2"}
dateparser = "^1.1.8"
spacy-alignments = "^0.9.1"
sentence-transformers = "^2.2.2"
uuid-utils = "^0.9.0"
orjson = "^3.9.12"
pybind11 = "^2.12.0"
fasttext-wheel = "^0.9.2"
lingua-language-detector = "^2.0.2"
transformers = "^4.45.2"
torch = "^2.4.1"
pydantic = "^2.9.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
sqlalchemy = "^2.0.35"
tqdm = "^4.66.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"
ipykernel = "^6.29.5"
jupyterlab = "^4.2.5"
ipywidgets = "^8.1.5"
tqdm = "^4.66.5"
matplotlib = "^3.9.2"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
python_files = "*_tests.py"

[tool.ruff]
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]
# Add "Q" to the list of enabled codes.

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
[tool.ruff.lint]
ignore = [
    "E501", # Never enforce `E501` (line length violations).
    "E701", # multiple statements on one line (colon) -  i use them for early returns
]
select = ["E", "F", "Q", "B"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
python_version = 3.8
warn_return_any = true
ignore_missing_imports = true
warn_unused_configs = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/noisemon/"
files = "$MYPY_CONFIG_FILE_DIR/noisemon/**/*.py"
exclude = "$MYPY_CONFIG_FILE_DIR/noisemon/**/*.tests.py"
python_executable = "$MYPY_CONFIG_FILE_DIR/.venv/bin/python"

